+++ src/tls/securetransport.cc
@@ -200,13 +200,13 @@ struct SecureTransportFilter : public pollster::Filter
    }
 
    void
-   Write(const void *buf, int len, const std::function<void(error*)> &onComplete)
+   Write(const void *buf, size_t len, const std::function<void(error*)> &onComplete)
    {
       Write(buf, len, onComplete, true);
    }
 
    void
-   Write(const void *buf, int len, const std::function<void(error*)> &onComplete, bool locked)
+   Write(const void *buf, size_t len, const std::function<void(error*)> &onComplete, bool locked)
    {
       std::shared_ptr<WriteCallback> writeCb;
       std::function<void(error*)> wrappedCb;
@@ -306,7 +306,7 @@ struct SecureTransportFilter : public pollster::Filter
    }
 
    void
-   OnBytesReceived(const void *buf, int len, error *err)
+   OnBytesReceived(const void *buf, size_t len, error *err)
    {
       currentReadBuf = buf;
       currentReadBufLen = len;
